6847 syscall header declaration incorrect Reviewed by: Yuri Pankov <yuri.pankov@nexenta.com> Reviewed by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net> Reviewed by: Garrett D'Amore <garrett@damore.org> Approved by: Dan McDonald <danmcd@omniti.com>

+++ usr/src/uts/common/sys/syscall.h
@@ -511,8 +511,8 @@ typedef struct {	/* return values from system call */
 
 #if !defined(_KERNEL)
 
-extern int	syscall(int, ...);
-extern int	__systemcall(sysret_t *, int, ...);
+extern long	syscall(int, ...);
+extern long	__systemcall(sysret_t *, int, ...);
 extern int	__set_errno(int);
 
 #endif	/* _KERNEL */
